Update installed RPMs before installing build dependencies in rpm.mk.
In some edge cases latest versions of build dependencies may conflict
with outdated versions of installed packages. The only solution is to
update the corresponding conflicting packages first, but RPM specs do
not allow that.

As a workaround, do a full system update before installing build
dependencies. This increases build times considerably, so making this
behavior optional may be the next step.