Skip non-exist packagecloud repository
It is necessary for packpack packaging using packpack itself. We have
the variables PACKAGECLOUD_USER and PACKAGECLOUD_REPO set in CI to use
in deploy, but packpack detects them during build and trying to enable
the corresponding repository. When the repository does not exist the
build fails.

So, to deploy a first package for a new distro version we need some
packages already being pushed to the repository (or pushed and then
deleted). It was very inconvenient to enable the new distro version
because of that.