Revert "Skip non-exist packagecloud repository"
This reverts commit f2f4d2d9660c97d2cb344126d5769ee3218bb72c.

Skipping of a repository lean on a fact that config file is not
available, but after packagecloud enables Ubuntu 18.10 Cosmic I realize
that the check does not work. The config file will be found, but than an
error about not found Release file will be fired during repositories
updating. See also [1]. The error is the following:

> E: The repository 'https://packagecloud.io/packpack/packpack/ubuntu cosmic Release' does not have a Release file.

Decided to remove this guessing at all, because it becomes complex and
unobvious.

[1]: https://github.com/packpack/packpack/issues/106